==> R/Seurat/analysis/scr/03_sample_vignette_SingleR.R <==
# AIM ---------------------------------------------------------------------
# sample snippet for the annotation of cell types using SingleR

# libraries ---------------------------------------------------------------
library(SingleR)
library(celldex)
library(Seurat)

# read in a sample seurat object ------------------------------------------
seu <- readRDS("../out/test_introns/object/data.combined_harmonySkipIntegration_AllSoupX_01000_06000_15_V5.rds")

# specify the version of Seurat Assay -------------------------------------
# set seurat compatible with seurat4 workflow
options(Seurat.object.assay.version = "v5")

# confirm the assay -------------------------------------------------------
class(seu@assays$RNA)

# sample workflow ---------------------------------------------------------
DimPlot(seu,raster = T,group.by = "RNA_snn_res.0.1")

# To do a fully automated annoation, we need a reference dataset of primary cells. 
# Any reference could be used. The package scRNAseq in Bioconductor includes several 
# scRNAseq datasets that can be used as reference to SingleR.

# CELLDEX DATABASES:

# BlueprintEncodeData 
# normalized expression values of 259 RNA-seq samples of pure stroma and immune cells as 
# generated and supplied by Blueprint and ENCODE

# DatabaseImmuneCellExpressionData
# normalized expression values of 1561 bulk RNA-seq samples of sorted cell populations 
# from the Database of Immune Cell Expression (DICE)

# HumanPrimaryCellAtlasData
# normalized expression values of the data stored in the Human Primary Cell Atlas.

# ImmGenData
# normalized expression values of 830 microarray samples of pure mouse immune cells, 
# generated by the Immunologic Genome Project (ImmGen)

# MonacoImmuneData
# normalized expression values of 114 bulk RNA-seq samples of sorted immune cell 
# populations that can be found in GSE107011

# MouseRNAseqData
# normalized expression values of 358 bulk RNA-seq samples of sorted cell populations 
# that can be found at GEO

# NovershternHematopoieticData
# normalized expression values of 211 bulk human microarray samples of sorted hematopoietic 
# cell populations that can be found in GSE24759
ref <- celldex::HumanPrimaryCellAtlasData()

# choose the level of cell type annotation
table(ref$label.main) # broad annotation
table(ref$label.fine) # fine-grained annotation that defines subtypes or states
table(ref$label.ont)  # fine-grained annotation mapped to the standard vocabulary in the Cell Ontology

# Now SingleR compares our normalized count data to a reference set, and finds the 
# most probable annation:
seu_annot <- SingleR(test = Seurat::GetAssayData(seu, layer = "data"),
                     ref = ref,
                     labels = ref$label.main)

# See what’s in there by using head:
head(seu_annot)

# There are some annotations that contain only a few cells. They are usually not 
# of interest, and they clogg our plots. Therefore we remove them from the annotation:
labels <- seu_annot$labels
t <- table(labels)
other <- names(t)[t < 10]
labels[labels %in% other] <- NA

# In order to visualize it in our UMAP, we have to add the annotation to 
# seu_int@meta.data:
seu$HPA_annot <- labels

# We can plot the annotations in the UMAP. Here, we use a different package for 
# plotting (dittoSeq) as it has a bit better default coloring, and some other 
# plotting functionality we will use later on.
library("dittoSeq")
dittoSeq::dittoDimPlot(seu, var = "HPA_annot", size = 1)
ggsave(filename = "../out/test_introns/plot/SingleR_annotation.png", 
       device = "png", width = 280, height = 210, units = "mm", dpi = 300)

rm(seu_annot)
rm(seu_ref)


# Other databases that can be used with SingleR:
# Most of them are available in the SingleCellExperiment format or they can be easily
# converted to be used with SingleR.


### 1. **Tabula Muris**
#    - **Descrizione**: Un'ampia collezione di dati di espressione genica su singola cellula proveniente da tessuti di topi a livello di singola cellula.
#    - **Utilità**: Ottimo per annotare tipi cellulari di tessuti murini. Contiene profili di molti tessuti diversi, rendendolo ideale per studi su modelli animali.
#    - **Link**: [Tabula Muris](https://tabula-muris.ds.czbiohub.org/)

### 2. **Tabula Sapiens**
#    - **Descrizione**: Un atlante completo delle cellule umane, costruito a partire da molteplici tessuti e organi. Offre una panoramica dettagliata dell'espressione genica nei principali tipi cellulari umani.
#    - **Utilità**: Perfetto per studi su una vasta gamma di tipi cellulari umani da tessuti diversi.
#    - **Link**: [Tabula Sapiens](https://tabula-sapiens-portal.ds.czbiohub.org/)

### 3. **HPA (Human Protein Atlas)**
#    - **Descrizione**: Il database contiene informazioni dettagliate sull'espressione genica e proteica a livello di singola cellula in tessuti umani. 
#    - **Utilità**: Ottimo per studi traslazionali su diversi tessuti e per annotare tipi cellulari umani basati su espressione genica e proteica.
#    - **Link**: [Human Protein Atlas](https://www.proteinatlas.org/)

### 4. **Mouse Cell Atlas (MCA)**
#    - **Descrizione**: Un atlante dell’espressione genica a singola cellula nei topi. Copre una vasta gamma di tipi cellulari e tessuti murini.
#    - **Utilità**: Ideale per annotazioni in studi su modelli murini, particolarmente per tessuti non immunitari.
#    - **Link**: [MCA](http://bis.zju.edu.cn/MCA/)

### 5. **PanglaoDB**
#    - **Descrizione**: Un database di dati di espressione genica a singola cellula provenienti da varie specie (umano e topo), con un forte focus su cellule staminali e progenitori.
#    - **Utilità**: Utile per annotare cellule staminali e progenitori sia umani che murini.
#    - **Link**: [PanglaoDB](https://panglaodb.se/)

### 6. **Allen Brain Atlas**
#    - **Descrizione**: Una risorsa di espressione genica per il cervello umano e murino, fornendo dati dettagliati per tipi cellulari specifici del cervello.
#    - **Utilità**: Perfetto per annotare cellule neuronali e studi neurobiologici.
#    - **Link**: [Allen Brain Atlas](https://portal.brain-map.org/)

### 7. **CellMarker**
#    - **Descrizione**: Un database di marker cellulari per la classificazione dei tipi di cellule umane e murine. Contiene marker specifici per una varietà di tipi cellulari e tessuti.
#    - **Utilità**: Utile per identificare tipi cellulari basati su marker noti piuttosto che su profili di espressione completi.
#    - **Link**: [CellMarker](http://biocc.hrbmu.edu.cn/CellMarker/)

### 8. **SCExpression**
#    - **Descrizione**: Un atlante online con dati di espressione genica a singola cellula da varie fonti, incluse più specie e tipi di tessuti.
#    - **Utilità**: Una risorsa completa per dati di espressione genica a singola cellula da più specie e tipi di tessuti.
#    - **Link**: [SCExpression](http://www.scexpression.org/)

### 9. **Single Cell Expression Atlas (EBI)**
#    - **Descrizione**: Un'ampia raccolta di dati di espressione genica a singola cellula da studi pubblicati, che copre molte specie e condizioni sperimentali.
#    - **Utilità**: Una risorsa per esplorare dati di espressione genica a singola cellula da più studi e specie.
#    - **Link**: [Single Cell Expression Atlas](https://www.ebi.ac.uk/gxa/sc/home)
